Move RewriteLink::alpha_convert to ScopeLink::alpha_convert
because it is gonna be useful to fix FreeVariables::substitute_scoped
to avoid name collision.
