+++ opencog/atoms/core/ScopeLink.cc
@@ -23,9 +23,8 @@
 
 #include <string>
 
-// #include <opencog/util/mt19937ar.h>
-// #include <opencog/util/random.h>
 // #include <opencog/util/Logger.h>
+#include <opencog/util/random.h>
 #include <opencog/atoms/atom_types/NameServer.h>
 #include <opencog/atoms/base/hash.h>
 #include <opencog/atoms/core/FindUtils.h>
@@ -149,6 +148,70 @@ void ScopeLink::init_scoped_variables(const Handle& vardecl)
 		HandleSeq owv(std::next(_outgoing.begin()), _outgoing.end());
 		_variables.canonical_sort(owv);
 	}
+}
+
+/* ================================================================= */
+
+inline Handle append_rand_str(const Handle& var)
+{
+	std::string new_var_name = randstr(var->get_name() + "-");
+	return createNode(var->get_type(), std::move(new_var_name));
+}
+
+inline HandleSeq append_rand_str(const HandleSeq& vars)
+{
+	HandleSeq new_vars;
+	for (const Handle& h : vars)
+		new_vars.push_back(append_rand_str(h));
+	return new_vars;
+}
+
+/**
+ * Wrap every glob node with a ListLink
+ *
+ * Since GlobNodes can be matched/substituted with one or more
+ * arguments, The arguments are expected to be wrapped with
+ * ListLink.
+ * In case of alpha conversion, Alpha converted GlobNodes in a
+ * program needs to be wrapped before passed to substitute the Glob.
+ */
+inline HandleSeq wrap_glob_with_list(const HandleSeq& vars)
+{
+	HandleSeq new_vars;
+	for (const Handle& var : vars) {
+		if (GLOB_NODE == var->get_type())
+			new_vars.push_back(createLink(HandleSeq{var}, LIST_LINK));
+		else new_vars.push_back(var);
+	}
+	return new_vars;
+}
+
+Handle ScopeLink::alpha_convert() const
+{
+	HandleSeq vars = append_rand_str(_variables.varseq);
+	return alpha_convert(vars);
+}
+
+Handle ScopeLink::alpha_convert(const HandleSeq& vars) const
+{
+	const auto wrapped = wrap_glob_with_list(vars);
+	// Perform alpha conversion
+	HandleSeq hs;
+	for (size_t i = 0; i < get_arity(); ++i)
+		hs.push_back(_variables.substitute_nocheck(getOutgoingAtom(i), wrapped));
+
+	// Create the alpha converted scope link
+	return createLink(std::move(hs), get_type());
+}
+
+Handle ScopeLink::alpha_convert(const HandleMap& vsmap) const
+{
+	HandleSeq vars;
+	for (const Handle& var : _variables.varseq) {
+		auto it = vsmap.find(var);
+		vars.push_back(it == vsmap.end() ? append_rand_str(var) : it->second);
+	}
+	return alpha_convert(vars);
 }
 
 /* ================================================================= */
